requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langsmith>=0.2.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
pytest>=8.0.0
